+++ common/src/main/java/me/rufia/fightorflight/CobblemonFightOrFlight.java
@@ -11,6 +11,7 @@ import me.rufia.fightorflight.goals.*;
 import me.rufia.fightorflight.net.CobblemonFightOrFlightNetwork;
 import me.rufia.fightorflight.utils.PokemonUtils;
 import me.rufia.fightorflight.utils.TargetingWhitelist;
+import me.rufia.fightorflight.utils.listeners.BehaviorDataListener;
 import me.rufia.fightorflight.utils.listeners.MoveDataListener;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
@@ -33,7 +34,7 @@ import java.util.Set;
 
 public class CobblemonFightOrFlight {
     public static final String MODID = "fightorflight";
-    public static final String COBBLEMON_MOD_ID="cobblemon";
+    public static final String COBBLEMON_MOD_ID = "cobblemon";
     public static final Logger LOGGER = LoggerFactory.getLogger(MODID);
     public static final float AUTO_AGGRO_THRESHOLD = 50.0f;
     private static FightOrFlightCommonConfigModel commonConfig;
@@ -63,6 +64,8 @@ public class CobblemonFightOrFlight {
         visualEffectConfig = AutoConfig.getConfigHolder(FightOrFlightVisualEffectConfigModel.class).getConfig();
         CobblemonFightOrFlightNetwork.init();
         ReloadListenerRegistry.register(PackType.SERVER_DATA, new MoveDataListener(), ResourceLocation.fromNamespaceAndPath(MODID, "movedata"));//unfinished
+        //ReloadListenerRegistry.register(PackType.SERVER_DATA,new BehaviorDataListener(),ResourceLocation.fromNamespaceAndPath(MODID, "behaviordata"));
+        //TODO enable this after the feature is finished
         TargetingWhitelist.init();
     }
 
@@ -109,7 +112,7 @@ public class CobblemonFightOrFlight {
         double natureAggressionCoefficient;
         double darknessAggressionCoefficient = 0;
         double intimidateCoefficient = 0;
-        String natureName = pokemon.getNature().getDisplayName().toLowerCase().replace("cobblemon.nature.", "");
+        String natureName = PokemonUtils.getNatureName(pokemon);
 
         if (Arrays.stream(CobblemonFightOrFlight.commonConfig().more_aggressive_nature).toList().contains(natureName)) {
             natureAggressionCoefficient = CobblemonFightOrFlight.commonConfig().more_aggressive_nature_multiplier;