+++ common/src/main/java/me/rufia/fightorflight/config/FightOrFlightVisualEffectConfigModel.java
@@ -27,16 +27,18 @@ public class FightOrFlightVisualEffectConfigModel implements ConfigData {
             "astralbarrage"
     };
     @Comment("Slicing moves")
-    public String[] slicing_moves={
-        "cut","razorleaf","slash","furycutter","aircutter","aerialace","leafblade","nightslash",
-            "airslash","x-scissor","psychocut","crosspoison","sacredsword","razorshell","secretsword",
-            "solarblade","behemothblade","stoneaxe","ceaselessedge","populationbomb","kowtowcleave",
-            "psyblade","bitterblade","aquacutter","mightycleave","tachyoncutter","spacialrend"
+    public String[] slicing_moves = {
+            "cut", "razorleaf", "slash", "furycutter", "aircutter", "aerialace", "leafblade", "nightslash",
+            "airslash", "x-scissor", "psychocut", "crosspoison", "sacredsword", "razorshell", "secretsword",
+            "solarblade", "behemothblade", "stoneaxe", "ceaselessedge", "populationbomb", "kowtowcleave",
+            "psyblade", "bitterblade", "aquacutter", "mightycleave", "tachyoncutter", "spacialrend"
     };
+    @Comment("If you want to enable the move indicator?")
+    public boolean enable_move_indicator = true;
     @Comment("Set the size of the move indicator")
-    public float move_indicator_size=0.5f;
+    public float move_indicator_size = 0.5f;
     @Comment("Set the horizontal position of the move indicator.(0-1,0 is the left side, 1 is the right side)")
-    public float move_indicator_x_relative =0.9f;
+    public float move_indicator_x_relative = 0.9f;
     @Comment("Set the horizontal position of the move indicator.(0-1,0 is the top, 1 is the bottom)")
-    public float move_indicator_y_relative =0.9f;
+    public float move_indicator_y_relative = 0.9f;
 }