+++ common/src/main/java/me/rufia/fightorflight/goals/PokemonPassiveAbilityGoal.java
@@ -8,7 +8,7 @@ import net.minecraft.world.entity.ai.goal.Goal;
 import net.minecraft.world.phys.AABB;
 
 import java.util.Objects;
-
+// UNUSED CURRENTLY
 //Some status moves are designed to be used passively,which means you just need to have it in the pokemon's moveset
 public class PokemonPassiveAbilityGoal extends Goal {
     protected PokemonEntity pokemonEntity;